+++ lib/influxdb/rails/railtie.rb
@@ -1,5 +1,5 @@
-require 'influxdb'
-require 'rails'
+require "influxdb"
+require "rails"
 
 module InfluxDB
   module Rails
@@ -8,21 +8,17 @@ module InfluxDB
         app.config.middleware.insert 0, InfluxDB::Rails::Rack
       end
 
-      # rubocop:disable Metrics/BlockLength
-
       config.after_initialize do
-        InfluxDB::Rails.configure(true) do |config|
-          config.load_rails_defaults
-        end
+        InfluxDB::Rails.configure(true, &:load_rails_defaults)
 
         ActiveSupport.on_load(:action_controller) do
-          require 'influxdb/rails/air_traffic_controller'
+          require "influxdb/rails/air_traffic_controller"
           include InfluxDB::Rails::AirTrafficController
-          require 'influxdb/rails/instrumentation'
+          require "influxdb/rails/instrumentation"
           include InfluxDB::Rails::Instrumentation
         end
 
-        require 'influxdb/rails/middleware/hijack_render_exception'
+        require "influxdb/rails/middleware/hijack_render_exception"
         ::ActionDispatch::DebugExceptions.prepend InfluxDB::Rails::Middleware::HijackRenderException
 
         if defined?(ActiveSupport::Notifications)
@@ -41,8 +37,6 @@ module InfluxDB
           ActiveSupport::Notifications.subscribe "process_action.action_controller", &listen
         end
       end
-
-      # rubocop:enable Metrics/BlockLength
     end
   end
 end