+++ spec/controllers/widgets_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require "spec_helper"
 
 RSpec.describe WidgetsController, type: :controller do
   describe "#new" do