+++ spec/unit/configuration_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require "spec_helper"
 
 RSpec.describe InfluxDB::Rails::Configuration do
   before do
@@ -81,28 +81,28 @@ RSpec.describe InfluxDB::Rails::Configuration do
 
   describe "#time_precision" do
     it "defaults to seconds" do
-      expect(InfluxDB::Rails.configuration.time_precision).to eql('s')
+      expect(InfluxDB::Rails.configuration.time_precision).to eql("s")
     end
 
     it "can be updated" do
       InfluxDB::Rails.configure do |config|
-        config.time_precision = 'ms'
+        config.time_precision = "ms"
       end
-      expect(InfluxDB::Rails.configuration.time_precision).to eql('ms')
+      expect(InfluxDB::Rails.configuration.time_precision).to eql("ms")
     end
   end
 
   describe "#rails_app_name" do
-    it 'defaults to nil' do
+    it "defaults to nil" do
       expect(InfluxDB::Rails.configuration.rails_app_name).to be(nil)
     end
 
     it "can be set to own name" do
       InfluxDB::Rails.configure do |config|
-        config.rails_app_name = 'my-app'
+        config.rails_app_name = "my-app"
       end
 
-      expect(InfluxDB::Rails.configuration.rails_app_name).to eq('my-app')
+      expect(InfluxDB::Rails.configuration.rails_app_name).to eq("my-app")
     end
   end
 end